# documents/openapi.yaml
openapi: 3.1.0
info:
  title: UdaConnect Refactored API
  version: 1.0.0
basePath: /
paths:
...
  /api/persons/{person_id}/connection:
    parameters:
      - in: query
        description: Proximity to a given user in meters
        name: distance
        schema:
          type: string
      - in: query
        description: Upper bound of date range
        name: end_date
        schema:
          type: string
      - in: query
        description: Lower bound of date range
        name: start_date
        schema:
          type: string
      - name: person_id
        in: path
        required: true
        schema:
          type: string
    get:
      responses:
        '200':
          description: Success
      operationId: get_connection_data_resource
      tags:
        - UdaConnect
...